// apps/backend/src/model/kanel/public/ServiceLink.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceInstanceId } from './ServiceInstance';

/** Identifier type for public.Service_Link */
export type ServiceLinkId = string & { __brand: 'public.Service_Link' };

/** Represents the table public.Service_Link */
export default interface ServiceLink {
  id: ServiceLinkId;

  service_instance_id: ServiceInstanceId | null;

  url: string | null;

  name: string | null;
}

/** Represents the initializer for the table public.Service_Link */
export interface ServiceLinkInitializer {
  /** Default value: gen_random_uuid() */
  id?: ServiceLinkId;

  service_instance_id?: ServiceInstanceId | null;

  url?: string | null;

  name?: string | null;
}

/** Represents the mutator for the table public.Service_Link */
export interface ServiceLinkMutator {
  id?: ServiceLinkId;

  service_instance_id?: ServiceInstanceId | null;

  url?: string | null;

  name?: string | null;
}